Structured_Output/requirement.txt
langchain-openai
python-dotenv
langchain_core